requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# 配置管理
python-dotenv==1.0.0

# 依赖注入
dependency-injector==4.41.0

# HTTP请求
requests==2.31.0

# 数据处理
numpy==1.24.3
pandas==2.0.3

# LLM客户端
openai==1.6.1

# 数据库（可选）
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# Redis缓存（可选）
# redis==5.0.1

# AI相关（按需安装）
# torch==2.1.0
# transformers==4.35.0

# 测试
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# 代码质量
# pylint==3.0.3
# black==23.12.1
# isort==5.13.2
# mypy==1.7.1
